+++ packages/snap/package.json
@@ -52,7 +52,7 @@
     "@types/qs": "^6.9.18",
     "@typescript-eslint/eslint-plugin": "^5.42.1",
     "@typescript-eslint/parser": "^5.42.1",
-    "bitcoindevkit": "^0.1.5",
+    "bitcoindevkit": "file:../../../bdk-wasm/pkg",
     "concurrently": "^9.1.0",
     "dotenv": "^16.4.5",
     "eslint": "^8.45.0",
