+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/snap-bitcoin-wallet.git"
   },
   "source": {
-    "shasum": "jofv9A/7+MQzhxPyCsINIxgiYXzQHusZINreUo7i1WI=",
+    "shasum": "vGsRf1ZAiPRofgPSv6dQTaVHxh4IunvZt4z231gEBrg=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
