+++ packages/snap/src/infra/jsx/components/index.ts
@@ -1,3 +1,4 @@
 export * from './SendForm';
 export * from './TransactionSummary';
 export * from './HeadingWithReturn';
+export * from './AssetIcon';
